Revert "Moved command and handler to right namespace"
This reverts commit 21e0ac652507d75fe6f38dedb3eac257fcc9cdf4.
