+++ HubOfChess.Application/PostLikes/Commands/CreatePostLikeCommand.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
 
-namespace HubOfChess.Application.PostLikes.Commands.CreatePostLike
+namespace HubOfChess.Application.PostLikes.Commands
 {
     public record CreatePostLikeCommand(Guid UserId, Guid PostId) : IRequest;
 }
